Make more Arc/Rc-compatible
Now `Arc` and `Rc` are typedefs for `Ash<'static, ...>`, so they
work as drop-in replacements for the `std` versions again.

Using other lifetime params now requires a different typedef,
with the placeholder names `ArcL` and `RcL`.

I will probably rename them to `Ash` and `Sh`, but need to
come up with a new name for the generic version first.